Исправления для сборки под Linux.

+++ LibrarySTD/ISCore/ISConstants.h
@@ -25,7 +25,7 @@ const char  CHAR_NULL_TERM = '\0'; //Символ конца строки
 //LOCALIZATION-----------------------------------------------------------------
 const char  LOCALIZATION_FILE_CARAT[] = "Carat";
 //DATABASE---------------------------------------------------------------------
-const size_t            DB_MAX_QUERY_TIME = 50; //Максимальное время (в милисекундах) выполнения SQL-запроса
+const long long         DB_MAX_QUERY_TIME = 50; //Максимальное время (в милисекундах) выполнения SQL-запроса
 const char              DB_CONNECTION_DEFAULT[] = "DefaultDB";
 //CONFIG-----------------------------------------------------------------------
 const char  CONFIG_TEMPLATE_SERVER[] = "Server"; //Шаблон сервера
